refactor: add hooks for product data
